backend/api/requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
psycopg2-binary
pydantic[email]
pwdlib[argon2]
pyjwt
python-multipart